refactor: root layout
