+++ src/app/(lang)/[lang]/layout.tsx
@@ -1,14 +1,7 @@
-import { setI18n } from "@lingui/react/server";
-import { Inter } from "next/font/google";
-import { Toaster } from "react-hot-toast";
-import { getI18nInstance, locales } from "~/i18n";
-import { LinguiClientProvider } from "~/providers/LinguiClientProvider";
-import { Providers } from "../../../providers";
-import "../../(main)/globals.css";
+import { locales } from "~/i18n";
+import { getRootLayout } from "../../_layout";
 
-const inter = Inter({ subsets: ["latin"] });
-
-export { metadata } from "../../(main)/layout";
+export { metadata } from "../../_layout";
 
 interface Props {
   params: Promise<{
@@ -21,22 +14,9 @@ export async function generateStaticParams() {
   return locales.map((locale) => ({ lang: locale }));
 }
 
-export default async function RootLayout({ params, children }: Props) {
+export default async function Layout({ params, children }: Props) {
   const { lang } = await params;
-  const i18n = getI18nInstance(lang);
-  setI18n(i18n);
+  const RootLayout = getRootLayout(lang);
 
-  return (
-    <html lang={lang}>
-      <body className={inter.className}>
-        <Toaster />
-        <LinguiClientProvider
-          initialLocale={lang}
-          initialMessages={i18n.messages}
-        >
-          <Providers>{children}</Providers>
-        </LinguiClientProvider>
-      </body>
-    </html>
-  );
+  return <RootLayout>{children}</RootLayout>;
 }
