+++ src/app/(main)/layout.tsx
@@ -1,37 +1,6 @@
-import { setI18n } from "@lingui/react/server";
-import { Inter } from "next/font/google";
-import { Toaster } from "react-hot-toast";
-import { getI18nInstance, sourceLocale } from "~/i18n";
-import { LinguiClientProvider } from "~/providers/LinguiClientProvider";
-import { Providers } from "../../providers";
-import "./globals.css";
-import type { Metadata } from "next";
-import type { ReactNode } from "react";
+import { getRootLayout } from "../_layout";
 
-const inter = Inter({ subsets: ["latin"] });
+export { metadata } from "../_layout";
 
-export const metadata: Metadata = {
-  title: "Magic tools",
-  description: "a magic tools for developers",
-  icons: ["/favicon.ico"],
-};
-
-export default function RootLayout({ children }: { children: ReactNode }) {
-  const lang = sourceLocale || "en";
-  const i18n = getI18nInstance(lang);
-  setI18n(i18n);
-
-  return (
-    <html lang={lang}>
-      <body className={inter.className}>
-        <Toaster />
-        <LinguiClientProvider
-          initialLocale={lang}
-          initialMessages={i18n.messages}
-        >
-          <Providers>{children}</Providers>
-        </LinguiClientProvider>
-      </body>
-    </html>
-  );
-}
+const RootLayout = getRootLayout("en");
+export default RootLayout;
